pyproject.toml
[tool.poetry]
name = "analyze-stocks"
version = "0.1.0"
description = ""
authors = ["Nitheesh-S <nitheeshmsk@gmail.com>"]
readme = "README.md"
packages = [{include = "analyze_stocks"}]

[tool.poetry.dependencies]
python = "^3.11"
fastapi = "^0.96.1"
sqlalchemy = "^2.0.16"
psycopg2-binary = "^2.9.6"
fyers-apiv2 = "^2.0.6"
hypercorn = "^0.14.3"
alembic = "^1.11.1"
black = "^23.3.0"
ipykernel = "^6.23.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
